pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ood-job-monitor"
version = "1.0.0"
description = "Open OnDemand utility for monitoring Slurm job efficiency and updating session cards"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
keywords = ["slurm", "hpc", "open-ondemand", "job-monitoring", "efficiency"]

dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
schmutz = "ood_job_monitor.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/ood-job-monitor"
Documentation = "https://github.com/yourusername/ood-job-monitor#readme"
Repository = "https://github.com/yourusername/ood-job-monitor"
Issues = "https://github.com/yourusername/ood-job-monitor/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ood_job_monitor = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ood_job_monitor --cov-report=term-missing"
